02.05 스토어 db 쿼리 dsl 작업중, textarea css 수정

+++ src/main/java/zerogreen/eco/dto/detail/LikesDto.java
@@ -20,7 +20,7 @@ public class LikesDto {
     private Long sno;
     private String storeName;
 
-    private int count;
+    private Long count;
 
     private StoreMember storeMember;
     private BasicUser basicUser;
